create Resource Card and move ForCombo init
